CHANGE ORGANIZATION OF REDUCERS
